requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.0
psycopg2-binary>=2.9.0
drf-yasg>=1.20.0
djangorestframework-simplejwt>=5.2.0
python-dotenv>=0.21.0
